requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0
werkzeug>=2.0.0
scipy>=1.8.0
gunicorn>=20.0.0
python-dotenv>=0.19.0
requests>=2.25.0
pymongo>=4.0.0
langchain>=0.0.1
groq>=0.3.0
openpyxl>=3.0.0
xlrd>=2.0.0
# Additional dependencies found in imports
SQLAlchemy>=1.4.0  # For backward compatibility
typing-extensions>=4.0.0
Pillow>=8.0.0  # For matplotlib
jsonschema>=4.0.0
setuptools>=58.0.0
wheel>=0.37.0
pytest>=6.0.0  # For testing